Create README.md
Created a folder and named it web-dev-lab, inside it i created an html file and named it index.html.
Next i copied and pasted the html code that was provided then opened the file using chrome. i re-configured 
git and navigated to the web-dev-lab folder and initialised git. Then i created a repository on git and connected
my local project to Github, later refreshed my Github and the index.html appeared, then i modified it by adding a
heading, a paragraph and an image.